commit my code

+++ gossip_bt_server.py
@@ -210,7 +210,7 @@ while True:
             c7 = raw * scale
 
 
-            SN3 =((c6 - 298) - ((0.03) * (c7 - 279))) / 0.276
+            SN3 =float(((c6 - 298) - ((0.03) * (c7 - 279))) / 0.276)
             SN3 = SN3 if (SN3 >= 0) else -SN3
             print("CO_SN3 : {}".format(SN3))
 
@@ -262,7 +262,7 @@ while True:
             c11 = (raw * scale)/1000
 
 
-            PM25 = (240.0*pow(c11,6) - 2491.3*pow(c11,5) + 9448.7*pow(c11,4) - 14840.0*pow(c11,3) + 10684.0*pow(c11,2) + 2211.8*c11 + 7.9623)
+            PM25 = float((240.0*pow(c11,6) - 2491.3*pow(c11,5) + 9448.7*pow(c11,4) - 14840.0*pow(c11,3) + 10684.0*pow(c11,2) + 2211.8*c11 + 7.9623))
             PM25 = PM25 if (SN4 >= 0) else -PM25
             print("PM25 : {}".format(PM25))
 
@@ -273,7 +273,7 @@ while True:
 
             def AQI_CALCULATION(Imin, Imax, Cmin, Cmax, Input):
                 Result = ((float(Imax) - float(Imin)) / (float(Cmax) - float(Cmin))) * (float(Input) - float(Cmin)) + float(Imin)
-                return Result
+                return float(Result)
 
 
             def O3(num):
